Move SeriesLegend to its own file
PiperOrigin-RevId: 215302632
