+++ charts_common/lib/common.dart
@@ -118,15 +118,11 @@ export 'src/chart/common/behavior/domain_highlighter.dart'
     show DomainHighlighter;
 export 'src/chart/common/behavior/initial_selection.dart' show InitialSelection;
 export 'src/chart/common/behavior/legend/legend.dart'
-    show
-        Legend,
-        LegendCellPadding,
-        LegendState,
-        LegendTapHandling,
-        SeriesLegend;
+    show Legend, LegendCellPadding, LegendState, LegendTapHandling;
 export 'src/chart/common/behavior/legend/legend_entry.dart' show LegendEntry;
 export 'src/chart/common/behavior/legend/legend_entry_generator.dart'
     show LegendEntryGenerator, LegendDefaultMeasure;
+export 'src/chart/common/behavior/legend/series_legend.dart' show SeriesLegend;
 export 'src/chart/common/behavior/line_point_highlighter.dart'
     show LinePointHighlighter, LinePointHighlighterFollowLineType;
 export 'src/chart/common/behavior/range_annotation.dart'
